Added missing variable 'print-weights' to the XfstCompiler constructor with impl argument. Added support for prolog format to hfst-fst2txt and a simple prolog test. Fixed a typo in configure.ac. Removed unnecessary commented text from HfstTransitionGraph.h.
git-svn-id: svn://svn.code.sf.net/p/hfst/code/trunk/hfst3@3554 941e2c2b-deac-454f-805a-451daa25f33c

+++ test/tools/txt2fst-functionality.sh
@@ -20,6 +20,12 @@ if ((test -z "$i") || $TOOLDIR/hfst-format --list-formats | grep $i > /dev/null)
         if ! $TOOLDIR/hfst-compare -s test cat$i  ; then
             exit 1
         fi
+        if ! $TOOLDIR/hfst-txt2fst --prolog $FFLAG $srcdir/cat.prolog > test ; then
+            exit 1
+        fi
+        if ! $TOOLDIR/hfst-compare -s test cat$i  ; then
+            exit 1
+        fi
         rm test
     fi
 fi
